requirements.txt
django-redis==5.4.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
requests==2.22.0
PyJWT==2.3.0
cryptography==3.4.7
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
redis==5.0.3
